requirements.txt
asgiref==3.3.1
boto3==1.16.33
botocore==1.19.33
django-crispy-forms==1.10.0
django-storages==1.10.1
jmespath==0.10.0
Pillow==8.0.1
python-dateutil==2.8.1
pytz==2020.4
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2